requirements.txt
ultralytics==8.1.10
opencv-python-headless==4.8.0.74
image_dehazer
Pillow